一部調整

+++ src/main/java/com/example/springboot_study/repository/StudentRepository.java
@@ -3,5 +3,8 @@ package com.example.springboot_study.repository;
 import com.example.springboot_study.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student,Long> {
+import java.util.List;
+
+public interface StudentRepository extends JpaRepository<Student, Integer> {
+
 }
